requirements.txt
torch
transformers
accelerate
vllm
torchdiffeq
datasets
hydra-core
omegaconf
wandb
